MOB-1674 - UI of checkout screen (#306)
* Created base class for collection card cell

* Implemented UI for 'get domain' card

* Improved card transition animation

* Added bookmarks to gitignore

* Added support of multiple sections in pull up view

* Added 'find new' option in top nav plus button

* Adjusted pull up height

* Added search domains view.
Added function to get domain suggestions

* Created purchase domains flow
Run purchase domains from entry points
Added handling of scroll offset from purchase search view

* Trim search query

* Added empty and error states to search

* Added icons

* Added checkout section

* Refactoring after code review

* Connected checkout view to purchase flow

* Code review updates

+++ unstoppable-ios-app/domains-manager-ios/Services/Firebase/Purchase/FirebasePurchaseEntities.swift
@@ -30,7 +30,7 @@ extension FirebasePurchaseDomainsService {
         let promoCreditsUsed: Int
         let storeCreditsUsed: Int
         let salesTax: Int
-        let taxRate: Int
+        let taxRate: Double
         let discounts: [DiscountDetails] // Assuming discounts are represented as strings
         let timestamp: Int
         @DecodeIgnoringFailed
